Keep QuantityaValue as the bounded version, introduce UnboundedQuantityValue as the new concept.
