+++ src/DataValues/QuantityValue.php
@@ -3,66 +3,66 @@
 namespace DataValues;
 
 use InvalidArgumentException;
-use LogicException;
 
 /**
- * Class representing a quantity with associated unit.
+ * Class representing a quantity with associated unit and uncertainty interval.
  * The amount is stored as a @see DecimalValue object.
  *
- * @see BoundedQuantityValue for quantities with a known uncertainty interval.
+ * @see UnboundedQuantityValue For quantities with no known uncertainty interval.
  * For simple numeric amounts use @see NumberValue.
  *
- * @note BoundedQuantityValue and plain QuantityValue both use the value type ID "quantity".
+ * @note UnboundedQuantityValue and QuantityValue both use the value type ID "quantity".
  * The fact that we use subclassing to model the bounded vs the unbounded case should be
  * considered an implementation detail.
  *
- * @since 0.1
+ * @since 0.7
  *
  * @license GPL-2.0+
  * @author Daniel Kinzler
  */
-class QuantityValue extends DataValueObject {
+class QuantityValue extends UnboundedQuantityValue {
 
 	/**
-	 * The quantity's amount
+	 * The quantity's upper bound
 	 *
 	 * @var DecimalValue
 	 */
-	protected $amount;
+	private $upperBound;
 
 	/**
-	 * The quantity's unit identifier (use "1" for unitless quantities).
+	 * The quantity's lower bound
 	 *
-	 * @var string
+	 * @var DecimalValue
 	 */
-	protected $unit;
+	private $lowerBound;
 
 	/**
 	 * @param DecimalValue $amount
 	 * @param string $unit A unit identifier. Must not be empty, use "1" for unit-less quantities.
+	 * @param DecimalValue $upperBound The upper bound of the quantity, inclusive.
+	 * @param DecimalValue $lowerBound The lower bound of the quantity, inclusive.
 	 *
 	 * @throws IllegalValueException
 	 */
-	public function __construct( DecimalValue $amount, $unit, $unused = 'nothing' ) {
-		if ( $unused !== 'nothing' ) {
-			throw new LogicException( 'Constructor called with old signature. '
-				. 'Perhaps you want to construct a BoundedQuantityValue instead.' );
+	public function __construct( DecimalValue $amount, $unit, DecimalValue $upperBound, DecimalValue $lowerBound ) {
+		parent::__construct( $amount, $unit );
+
+		if ( $lowerBound->compare( $amount ) > 0 ) {
+			throw new IllegalValueException( '$lowerBound ' . $lowerBound->getValue() . ' must be <= $amount ' . $amount->getValue() );
 		}
 
-		if ( !is_string( $unit ) ) {
-			throw new IllegalValueException( '$unit needs to be a string, not ' . gettype( $unit ) );
+		if ( $upperBound->compare( $amount ) < 0 ) {
+			throw new IllegalValueException( '$upperBound ' . $upperBound->getValue() . ' must be >= $amount ' . $amount->getValue() );
 		}
 
-		if ( $unit === '' ) {
-			throw new IllegalValueException( '$unit can not be an empty string (use "1" for unit-less quantities)' );
-		}
-
-		$this->amount = $amount;
-		$this->unit = $unit;
+		$this->upperBound = $upperBound;
+		$this->lowerBound = $lowerBound;
 	}
 
 	/**
 	 * Returns a QuantityValue representing the given amount.
+	 * If no upper or lower bound is given, the amount is assumed to be absolutely exact,
+	 * that is, the amount itself will be used as the upper and lower bound.
 	 *
 	 * This is a convenience wrapper around the constructor that accepts native values
 	 * instead of DecimalValue objects.
@@ -70,18 +70,22 @@ class QuantityValue extends DataValueObject {
 	 * @note: if the amount or a bound is given as a string, the string must conform
 	 * to the rules defined by @see DecimalValue.
 	 *
-	 * @since 0.1
+	 * @since 0.7
 	 *
-	 * @param string|int|float|DecimalValue $number
+	 * @param string|int|float|DecimalValue|UnboundedQuantityValue $number
 	 * @param string $unit A unit identifier. Must not be empty, use "1" for unit-less quantities.
+	 * @param string|int|float|DecimalValue|null $upperBound
+	 * @param string|int|float|DecimalValue|null $lowerBound
 	 *
-	 * @return self
+	 * @return QuantityValue
 	 * @throws IllegalValueException
 	 */
-	public static function newFromNumber( $number, $unit = '1' ) {
+	public static function newFromNumber( $number, $unit = '1', $upperBound = null, $lowerBound = null ) {
 		$number = self::asDecimalValue( 'amount', $number );
+		$upperBound = self::asDecimalValue( 'upperBound', $upperBound, $number );
+		$lowerBound = self::asDecimalValue( 'lowerBound', $lowerBound, $number );
 
-		return new self( $number, $unit );
+		return new self( $number, $unit, $upperBound, $lowerBound );
 	}
 
 	/**
@@ -89,134 +93,130 @@ class QuantityValue extends DataValueObject {
 	 *
 	 * @deprecated since 0.1, use newFromNumber instead
 	 *
-	 * @param string|int|float|DecimalValue $number
+	 * @param string|int|float|DecimalValue|UnboundedQuantityValue $number
 	 * @param string $unit
 	 * @param string|int|float|DecimalValue|null $upperBound
 	 * @param string|int|float|DecimalValue|null $lowerBound
 	 *
-	 * @return self
+	 * @return QuantityValue
 	 */
-	public static function newFromDecimal( $number, $unit = '1' ) {
-		return self::newFromNumber( $number, $unit );
-	}
-
-	/**
-	 * Converts $number to a DecimalValue if possible and necessary.
-	 *
-	 * @note: if the $number is given as a string, it must conform to the rules
-	 *        defined by @see DecimalValue.
-	 *
-	 * @param string $name The variable name to use in exception messages
-	 * @param string|int|float|DecimalValue|null $number
-	 * @param DecimalValue|null $default
-	 *
-	 * @throws IllegalValueException
-	 * @throws InvalidArgumentException
-	 * @return DecimalValue
-	 */
-	protected static function asDecimalValue( $name, $number, DecimalValue $default = null ) {
-		if ( !is_string( $name ) ) {
-			throw new InvalidArgumentException( '$name must be a string' );
-		}
-
-		if ( $number === null ) {
-			if ( $default === null ) {
-				throw new InvalidArgumentException( '$' . $name . ' must not be null' );
-			}
-
-			$number = $default;
-		}
-
-		if ( $number instanceof DecimalValue ) {
-			// nothing to do
-		} elseif ( is_int( $number ) || is_float( $number ) || is_string( $number ) ) {
-			$number = new DecimalValue( $number );
-		} else {
-			throw new IllegalValueException( '$' . $name . '  must be a string, int, or float' );
-		}
-
-		return $number;
-	}
-
-	/**
-	 * @see Serializable::serialize
-	 *
-	 * @since 0.1
-	 *
-	 * @return string
-	 */
-	public function serialize() {
-		return serialize( array_values( $this->getArrayValue() ) );
+	public static function newFromDecimal( $number, $unit = '1', $upperBound = null, $lowerBound = null ) {
+		return self::newFromNumber( $number, $unit, $upperBound, $lowerBound );
 	}
 
 	/**
 	 * @see Serializable::unserialize
 	 *
-	 * @since 0.1
+	 * @since 0.7
 	 *
 	 * @param string $data
 	 */
 	public function unserialize( $data ) {
-		list( $amount, $unit ) = unserialize( $data );
+		list( $amount, $unit, $upperBound, $lowerBound ) = unserialize( $data );
 		$amount = DecimalValue::newFromArray( $amount );
-		$this->__construct( $amount, $unit );
+		$upperBound = DecimalValue::newFromArray( $upperBound );
+		$lowerBound = DecimalValue::newFromArray( $lowerBound );
+		$this->__construct( $amount, $unit, $upperBound, $lowerBound );
 	}
 
 	/**
-	 * @see DataValue::getType
+	 * Returns this quantity's upper bound.
 	 *
-	 * @since 0.1
-	 *
-	 * @return string
-	 */
-	public static function getType() {
-		return 'quantity';
-	}
-
-	/**
-	 * @see DataValue::getSortKey
-	 *
-	 * @since 0.1
-	 *
-	 * @return float
-	 */
-	public function getSortKey() {
-		return $this->amount->getValueFloat();
-	}
-
-	/**
-	 * Returns the quantity object.
-	 * @see DataValue::getValue
-	 *
-	 * @since 0.1
-	 *
-	 * @return self
-	 */
-	public function getValue() {
-		return $this;
-	}
-
-	/**
-	 * Returns the amount represented by this quantity.
-	 *
-	 * @since 0.1
+	 * @since 0.7
 	 *
 	 * @return DecimalValue
 	 */
-	public function getAmount() {
-		return $this->amount;
+	public function getUpperBound() {
+		return $this->upperBound;
 	}
 
 	/**
-	 * Returns the unit held by this quantity.
-	 * Unit-less quantities should use "1" as their unit.
+	 * Returns this quantity's lower bound.
 	 *
-	 * @since 0.1
+	 * @since 0.7
 	 *
-	 * @return string
+	 * @return DecimalValue
 	 */
-	public function getUnit() {
-		return $this->unit;
+	public function getLowerBound() {
+		return $this->lowerBound;
+	}
+
+	/**
+	 * Returns the size of the uncertainty interval.
+	 * This can roughly be interpreted as "amount +/- uncertainty/2".
+	 *
+	 * The exact interpretation of the uncertainty interval is left to the concrete application or
+	 * data point. For example, the uncertainty interval may be defined to be that part of a
+	 * normal distribution that is required to cover the 95th percentile.
+	 *
+	 * @since 0.7
+	 *
+	 * @return float
+	 */
+	public function getUncertainty() {
+		return $this->upperBound->getValueFloat() - $this->lowerBound->getValueFloat();
+	}
+
+	/**
+	 * Returns a DecimalValue representing the symmetrical offset to be applied
+	 * to the raw amount for a rough representation of the uncertainty interval,
+	 * as in "amount +/- offset".
+	 *
+	 * The offset is calculated as max( amount - lowerBound, upperBound - amount ).
+	 *
+	 * @since 0.7
+	 *
+	 * @return DecimalValue
+	 */
+	public function getUncertaintyMargin() {
+		$math = new DecimalMath();
+
+		$lowerMargin = $math->sum( $this->amount, $this->lowerBound->computeComplement() );
+		$upperMargin = $math->sum( $this->upperBound, $this->amount->computeComplement() );
+
+		$margin = $math->max( $lowerMargin, $upperMargin );
+		return $margin;
+	}
+
+	/**
+	 * Returns the order of magnitude of the uncertainty as the exponent of
+	 * last significant digit in the amount-string. The value returned by this
+	 * is suitable for use with @see DecimalMath::roundToExponent().
+	 *
+	 * @example: if two digits after the decimal point are significant, this
+	 * returns -2.
+	 *
+	 * @example: if the last two digits before the decimal point are insignificant,
+	 * this returns 2.
+	 *
+	 * Note that this calculation assumes a symmetric uncertainty interval,
+	 * and can be misleading.
+	 *
+	 * @since 0.7
+	 *
+	 * @return int
+	 */
+	public function getOrderOfUncertainty() {
+		// the desired precision is given by the distance between the amount and
+		// whatever is closer, the upper or lower bound.
+		//TODO: use DecimalMath to avoid floating point errors!
+		$amount = $this->amount->getValueFloat();
+		$upperBound = $this->upperBound->getValueFloat();
+		$lowerBound = $this->lowerBound->getValueFloat();
+		$precision = min( $amount - $lowerBound, $upperBound - $amount );
+
+		if ( $precision === 0.0 ) {
+			// If there is no uncertainty, the order of uncertainty is a bit more than what we have digits for.
+			return -strlen( $this->amount->getFractionalPart() );
+		}
+
+		// e.g. +/- 200 -> 2; +/- 0.02 -> -2
+		// note: we really want floor( log10( $precision ) ), but have to account for
+		// small errors made in the floating point operations above.
+		// @todo: use bcmath (via DecimalMath) to avoid this if possible
+		$orderOfUncertainty = floor( log10( $precision + 0.0000000005 ) );
+
+		return (int)$orderOfUncertainty;
 	}
 
 	/**
@@ -244,7 +244,7 @@ class QuantityValue extends DataValueObject {
 	 * @param mixed ... Any extra parameters will be passed to the $transformation function.
 	 *
 	 * @throws InvalidArgumentException
-	 * @return self
+	 * @return QuantityValue
 	 */
 	public function transform( $newUnit, $transformation ) {
 		if ( !is_callable( $transformation ) ) {
@@ -274,19 +274,38 @@ class QuantityValue extends DataValueObject {
 		$args[0] = $this->amount;
 		$amount = call_user_func_array( $transformation, $args );
 
+		$args[0] = $this->upperBound;
+		$upperBound = call_user_func_array( $transformation, $args );
+
+		$args[0] = $this->lowerBound;
+		$lowerBound = call_user_func_array( $transformation, $args );
+
 		// use a preliminary QuantityValue to determine the significant number of digits
-		return new self( $amount, $newUnit );
+		$transformed = new self( $amount, $newUnit, $upperBound, $lowerBound );
+		$roundingExponent = $transformed->getOrderOfUncertainty();
+
+		// apply rounding to the significant digits
+		$math = new DecimalMath();
+
+		$amount = $math->roundToExponent( $amount, $roundingExponent );
+		$upperBound = $math->roundToExponent( $upperBound, $roundingExponent );
+		$lowerBound = $math->roundToExponent( $lowerBound, $roundingExponent );
+
+		return new self( $amount, $newUnit, $upperBound, $lowerBound );
 	}
 
 	public function __toString() {
 		return $this->amount->getValue()
+			. '[' . $this->lowerBound->getValue()
+			. '..' . $this->upperBound->getValue()
+			. ']'
 			. ( $this->unit === '1' ? '' : $this->unit );
 	}
 
 	/**
 	 * @see DataValue::getArrayValue
 	 *
-	 * @since 0.1
+	 * @since 0.7
 	 *
 	 * @return array
 	 */
@@ -294,6 +313,8 @@ class QuantityValue extends DataValueObject {
 		return array(
 			'amount' => $this->amount->getArrayValue(),
 			'unit' => $this->unit,
+			'upperBound' => $this->upperBound->getArrayValue(),
+			'lowerBound' => $this->lowerBound->getArrayValue(),
 		);
 	}
 
@@ -301,38 +322,27 @@ class QuantityValue extends DataValueObject {
 	 * Constructs a new instance of the DataValue from the provided data.
 	 * This can round-trip with @see getArrayValue
 	 *
-	 * @since 0.1
+	 * @since 0.7
 	 *
 	 * @param mixed $data
 	 *
-	 * @return self
+	 * @return QuantityValue
 	 * @throws IllegalValueException
 	 */
 	public static function newFromArray( $data ) {
-		self::requireArrayFields( $data, array( 'amount', 'unit' ) );
+		if ( !isset( $data['upperBound'] ) && isset( $data['lowerBound'] ) ) {
+			// No bounds given, so construct an unbounded QuantityValue.
+			return parent::newFromArray( $data );
+		}
+
+		self::requireArrayFields( $data, array( 'amount', 'unit', 'upperBound', 'lowerBound' ) );
 
 		return new static(
 			DecimalValue::newFromArray( $data['amount'] ),
-			$data['unit']
+			$data['unit'],
+			DecimalValue::newFromArray( $data['upperBound'] ),
+			DecimalValue::newFromArray( $data['lowerBound'] )
 		);
-	}
-
-	/**
-	 * @see Comparable::equals
-	 *
-	 * @since 0.1
-	 *
-	 * @param mixed $target
-	 *
-	 * @return bool
-	 */
-	public function equals( $target ) {
-		if ( $this === $target ) {
-			return true;
-		}
-
-		return $target instanceof self
-			&& $this->toArray() === $target->toArray();
 	}
 
 }
