+++ tests/ValueFormatters/QuantityFormatterTest.php
@@ -2,8 +2,8 @@
 
 namespace ValueFormatters\Test;
 
-use DataValues\BoundedQuantityValue;
 use DataValues\QuantityValue;
+use DataValues\UnboundedQuantityValue;
 use ValueFormatters\DecimalFormatter;
 use ValueFormatters\FormatterOptions;
 use ValueFormatters\QuantityFormatter;
@@ -104,45 +104,45 @@ class QuantityFormatterTest extends ValueFormatterTestBase {
 		) );
 
 		return array(
-			'+0/nm' => array( BoundedQuantityValue::newFromNumber( '+0', '1', '+0', '+0' ), '0', $noMargin ),
-			'+0/wm' => array( BoundedQuantityValue::newFromNumber( '+0', '1', '+0', '+0' ), '0±0', $withMargin ),
-			'+0/zm' => array( BoundedQuantityValue::newFromNumber( '+0', '1', '+0', '+0' ), '0', $withNonZeroMargin ),
+			'+0/nm' => array( QuantityValue::newFromNumber( '+0', '1', '+0', '+0' ), '0', $noMargin ),
+			'+0/wm' => array( QuantityValue::newFromNumber( '+0', '1', '+0', '+0' ), '0±0', $withMargin ),
+			'+0/zm' => array( QuantityValue::newFromNumber( '+0', '1', '+0', '+0' ), '0', $withNonZeroMargin ),
 
-			'+0.0/nm' => array( BoundedQuantityValue::newFromNumber( '+0.0', '°', '+0.1', '-0.1' ), '0.0 °', $noMargin ),
-			'+0.0/wm' => array( BoundedQuantityValue::newFromNumber( '+0.0', '°', '+0.1', '-0.1' ), '0.0±0.1 °', $withMargin ),
-			'+0.0/xr' => array( BoundedQuantityValue::newFromNumber( '+0.0', '°', '+0.1', '-0.1' ), '0.00±0.10 °', $exactRounding ),
+			'+0.0/nm' => array( QuantityValue::newFromNumber( '+0.0', '°', '+0.1', '-0.1' ), '0.0 °', $noMargin ),
+			'+0.0/wm' => array( QuantityValue::newFromNumber( '+0.0', '°', '+0.1', '-0.1' ), '0.0±0.1 °', $withMargin ),
+			'+0.0/xr' => array( QuantityValue::newFromNumber( '+0.0', '°', '+0.1', '-0.1' ), '0.00±0.10 °', $exactRounding ),
 
-			'-1205/nm' => array( BoundedQuantityValue::newFromNumber( '-1205', 'm', '-1105', '-1305' ), '-1200 m', $noMargin ),
-			'-1205/wm' => array( BoundedQuantityValue::newFromNumber( '-1205', 'm', '-1105', '-1305' ), '-1200±100 m', $withMargin ),
-			'-1205/nr' => array( BoundedQuantityValue::newFromNumber( '-1205', 'm', '-1105', '-1305' ), '-1205±100 m', $noRounding ),
-			'-1205/xr' => array( BoundedQuantityValue::newFromNumber( '-1205', 'm', '-1105', '-1305' ), '-1205.00±100.00 m', $exactRounding ),
-			'-1205/nu' => array( BoundedQuantityValue::newFromNumber( '-1205', 'm', '-1105', '-1305' ), '-1200±100', $noUnit ),
+			'-1205/nm' => array( QuantityValue::newFromNumber( '-1205', 'm', '-1105', '-1305' ), '-1200 m', $noMargin ),
+			'-1205/wm' => array( QuantityValue::newFromNumber( '-1205', 'm', '-1105', '-1305' ), '-1200±100 m', $withMargin ),
+			'-1205/nr' => array( QuantityValue::newFromNumber( '-1205', 'm', '-1105', '-1305' ), '-1205±100 m', $noRounding ),
+			'-1205/xr' => array( QuantityValue::newFromNumber( '-1205', 'm', '-1105', '-1305' ), '-1205.00±100.00 m', $exactRounding ),
+			'-1205/nu' => array( QuantityValue::newFromNumber( '-1205', 'm', '-1105', '-1305' ), '-1200±100', $noUnit ),
 
-			'+3.025/nm' => array( BoundedQuantityValue::newFromNumber( '+3.025', '1', '+3.02744', '+3.0211' ), '3.025', $noMargin ),
-			'+3.025/wm' => array( BoundedQuantityValue::newFromNumber( '+3.025', '1', '+3.02744', '+3.0211' ), '3.025±0.004', $withMargin ),
-			'+3.025/zm' => array( BoundedQuantityValue::newFromNumber( '+3.025', '1', '+3.02744', '+3.0211' ), '3.025±0.004', $withNonZeroMargin ),
-			'+3.025/xr' => array( BoundedQuantityValue::newFromNumber( '+3.025', '1', '+3.02744', '+3.0211' ), '3.03±0.00', $exactRounding ), // TODO: never round to 0! See bug #56892
+			'+3.025/nm' => array( QuantityValue::newFromNumber( '+3.025', '1', '+3.02744', '+3.0211' ), '3.025', $noMargin ),
+			'+3.025/wm' => array( QuantityValue::newFromNumber( '+3.025', '1', '+3.02744', '+3.0211' ), '3.025±0.004', $withMargin ),
+			'+3.025/zm' => array( QuantityValue::newFromNumber( '+3.025', '1', '+3.02744', '+3.0211' ), '3.025±0.004', $withNonZeroMargin ),
+			'+3.025/xr' => array( QuantityValue::newFromNumber( '+3.025', '1', '+3.02744', '+3.0211' ), '3.03±0.00', $exactRounding ), // TODO: never round to 0! See bug #56892
 
-			'+3.125/nr' => array( BoundedQuantityValue::newFromNumber( '+3.125', '1', '+3.2', '+3.0' ), '3.125±0.125', $noRounding ),
-			'+3.125/xr' => array( BoundedQuantityValue::newFromNumber( '+3.125', '1', '+3.2', '+3.0' ), '3.13±0.13', $exactRounding ),
+			'+3.125/nr' => array( QuantityValue::newFromNumber( '+3.125', '1', '+3.2', '+3.0' ), '3.125±0.125', $noRounding ),
+			'+3.125/xr' => array( QuantityValue::newFromNumber( '+3.125', '1', '+3.2', '+3.0' ), '3.13±0.13', $exactRounding ),
 
-			'+3.125/fs' => array( BoundedQuantityValue::newFromNumber( '+3.125', '1', '+3.2', '+3.0' ), '+3.13', $forceSign ),
+			'+3.125/fs' => array( QuantityValue::newFromNumber( '+3.125', '1', '+3.2', '+3.0' ), '+3.13', $forceSign ),
 
-			'UB: +0.0/nm' => array( QuantityValue::newFromNumber( '+0.0', '°' ), '0.0 °', $noMargin ),
-			'UB: +0.0/wm' => array( QuantityValue::newFromNumber( '+0.0', '°' ), '0.0 °', $withMargin ),
-			'UB: +0.0/zm' => array( QuantityValue::newFromNumber( '+0.0', '°' ), '0.0 °', $withNonZeroMargin ),
-			'UB: +0.0/xr' => array( QuantityValue::newFromNumber( '+0.0', '°' ), '0.00 °', $exactRounding ),
-			'UB: +5.020/nm' => array( QuantityValue::newFromNumber( '+5.020', '°' ), '5.020 °', $noMargin ),
-			'UB: +5.020/wm' => array( QuantityValue::newFromNumber( '+5.020', '°' ), '5.020 °', $withMargin ),
-			'UB: +5.020/zm' => array( QuantityValue::newFromNumber( '+5.020', '°' ), '5.020 °', $withNonZeroMargin ),
-			'UB: +5.020/xr' => array( QuantityValue::newFromNumber( '+5.020', '°' ), '5.02 °', $exactRounding ),
-			'UB: +3.125/fs' => array( QuantityValue::newFromNumber( '+3.125', '1' ), '+3.125', $forceSign ),
+			'UB: +0.0/nm' => array( UnboundedQuantityValue::newFromNumber( '+0.0', '°' ), '0.0 °', $noMargin ),
+			'UB: +0.0/wm' => array( UnboundedQuantityValue::newFromNumber( '+0.0', '°' ), '0.0 °', $withMargin ),
+			'UB: +0.0/zm' => array( UnboundedQuantityValue::newFromNumber( '+0.0', '°' ), '0.0 °', $withNonZeroMargin ),
+			'UB: +0.0/xr' => array( UnboundedQuantityValue::newFromNumber( '+0.0', '°' ), '0.00 °', $exactRounding ),
+			'UB: +5.020/nm' => array( UnboundedQuantityValue::newFromNumber( '+5.020', '°' ), '5.020 °', $noMargin ),
+			'UB: +5.020/wm' => array( UnboundedQuantityValue::newFromNumber( '+5.020', '°' ), '5.020 °', $withMargin ),
+			'UB: +5.020/zm' => array( UnboundedQuantityValue::newFromNumber( '+5.020', '°' ), '5.020 °', $withNonZeroMargin ),
+			'UB: +5.020/xr' => array( UnboundedQuantityValue::newFromNumber( '+5.020', '°' ), '5.02 °', $exactRounding ),
+			'UB: +3.125/fs' => array( UnboundedQuantityValue::newFromNumber( '+3.125', '1' ), '+3.125', $forceSign ),
 		);
 	}
 
 	public function testFormatWithFormatString() {
 		$formatter = $this->getQuantityFormatter( null, '<$2>$1' );
-		$value = QuantityValue::newFromNumber( '+5', 'USD' );
+		$value = UnboundedQuantityValue::newFromNumber( '+5', 'USD' );
 		$formatted = $formatter->format( $value );
 		$this->assertSame( '<USD>5', $formatted );
 	}
