+++ tests/ValueFormatters/QuantityHtmlFormatterTest.php
@@ -2,8 +2,8 @@
 
 namespace ValueFormatters\Test;
 
-use DataValues\BoundedQuantityValue;
 use DataValues\QuantityValue;
+use DataValues\UnboundedQuantityValue;
 use ValueFormatters\DecimalFormatter;
 use ValueFormatters\FormatterOptions;
 use ValueFormatters\QuantityHtmlFormatter;
@@ -67,23 +67,23 @@ class QuantityHtmlFormatterTest extends ValueFormatterTestBase {
 	public function validProvider() {
 		return array(
 			'Unbounded, Unit 1' => array(
-				QuantityValue::newFromNumber( '+2', '1' ),
+				UnboundedQuantityValue::newFromNumber( '+2', '1' ),
 				'2'
 			),
 			'Unbounded, String unit' => array(
-				QuantityValue::newFromNumber( '+2', 'Ultrameter' ),
+				UnboundedQuantityValue::newFromNumber( '+2', 'Ultrameter' ),
 				'2 <span class="wb-unit">Ultrameter</span>'
 			),
 			'Bounded, Unit 1' => array(
-				BoundedQuantityValue::newFromNumber( '+2', '1', '+3', '+1' ),
+				QuantityValue::newFromNumber( '+2', '1', '+3', '+1' ),
 				'2±1'
 			),
 			'Bounded, String unit' => array(
-				BoundedQuantityValue::newFromNumber( '+2', 'Ultrameter', '+3', '+1' ),
+				QuantityValue::newFromNumber( '+2', 'Ultrameter', '+3', '+1' ),
 				'2±1 <span class="wb-unit">Ultrameter</span>'
 			),
 			'HTML injection' => array(
-				BoundedQuantityValue::newFromNumber( '+2', '<b>injection</b>', '+2', '+2' ),
+				QuantityValue::newFromNumber( '+2', '<b>injection</b>', '+2', '+2' ),
 				'2±0 <span class="wb-unit">&lt;b&gt;injection&lt;/b&gt;</span>'
 			),
 		);
@@ -91,7 +91,7 @@ class QuantityHtmlFormatterTest extends ValueFormatterTestBase {
 
 	public function testFormatWithFormatString() {
 		$formatter = $this->getQuantityHtmlFormatter( null, null, '$2&thinsp;$1' );
-		$value = QuantityValue::newFromNumber( '+5', 'USD' );
+		$value = UnboundedQuantityValue::newFromNumber( '+5', 'USD' );
 		$formatted = $formatter->format( $value );
 		$this->assertSame( '<span class="wb-unit">USD</span>&thinsp;5', $formatted );
 	}
@@ -110,7 +110,7 @@ class QuantityHtmlFormatterTest extends ValueFormatterTestBase {
 			->will( $this->returnValue( '<b>+2</b>' ) );
 
 		$formatter = $this->getQuantityHtmlFormatter( $options, $decimalFormatter );
-		$formatted = $formatter->format( QuantityValue::newFromNumber( '+2', $unit ) );
+		$formatted = $formatter->format( UnboundedQuantityValue::newFromNumber( '+2', $unit ) );
 		$this->assertSame( $expected, $formatted );
 	}
