+++ tests/ValueParsers/QuantityParserTest.php
@@ -2,8 +2,8 @@
 
 namespace ValueParsers\Test;
 
-use DataValues\BoundedQuantityValue;
 use DataValues\QuantityValue;
+use DataValues\UnboundedQuantityValue;
 use ValueParsers\ParserOptions;
 use ValueParsers\QuantityParser;
 use ValueParsers\ValueParser;
@@ -66,68 +66,68 @@ class QuantityParserTest extends StringValueParserTest {
 	public function validInputProvider() {
 		$amounts = array(
 			// amounts in various styles and forms
-			'0' => QuantityValue::newFromNumber( 0, '1' ),
-			'-0' => QuantityValue::newFromNumber( 0, '1' ),
-			'-00.00' => QuantityValue::newFromNumber( '+0.00', '1' ),
-			'+00.00' => QuantityValue::newFromNumber( '+0.00', '1' ),
-			'0001' => QuantityValue::newFromNumber( 1, '1' ),
-			'+01' => QuantityValue::newFromNumber( 1, '1' ),
-			'-1' => QuantityValue::newFromNumber( -1, '1' ),
-			'+42' => QuantityValue::newFromNumber( 42, '1' ),
-			' -  42' => QuantityValue::newFromNumber( -42, '1' ),
-			'9001' => QuantityValue::newFromNumber( 9001, '1' ),
-			'.5' => QuantityValue::newFromNumber( '+0.5', '1' ),
-			'-.125' => QuantityValue::newFromNumber( '-0.125', '1' ),
-			'3.' => QuantityValue::newFromNumber( 3, '1' ),
-			' 3 ' => QuantityValue::newFromNumber( 3, '1' ),
-			'2.125' => QuantityValue::newFromNumber( '+2.125', '1' ),
-			'2.1250' => QuantityValue::newFromNumber( '+2.1250', '1' ),
+			'0' => UnboundedQuantityValue::newFromNumber( 0, '1' ),
+			'-0' => UnboundedQuantityValue::newFromNumber( 0, '1' ),
+			'-00.00' => UnboundedQuantityValue::newFromNumber( '+0.00', '1' ),
+			'+00.00' => UnboundedQuantityValue::newFromNumber( '+0.00', '1' ),
+			'0001' => UnboundedQuantityValue::newFromNumber( 1, '1' ),
+			'+01' => UnboundedQuantityValue::newFromNumber( 1, '1' ),
+			'-1' => UnboundedQuantityValue::newFromNumber( -1, '1' ),
+			'+42' => UnboundedQuantityValue::newFromNumber( 42, '1' ),
+			' -  42' => UnboundedQuantityValue::newFromNumber( -42, '1' ),
+			'9001' => UnboundedQuantityValue::newFromNumber( 9001, '1' ),
+			'.5' => UnboundedQuantityValue::newFromNumber( '+0.5', '1' ),
+			'-.125' => UnboundedQuantityValue::newFromNumber( '-0.125', '1' ),
+			'3.' => UnboundedQuantityValue::newFromNumber( 3, '1' ),
+			' 3 ' => UnboundedQuantityValue::newFromNumber( 3, '1' ),
+			'2.125' => UnboundedQuantityValue::newFromNumber( '+2.125', '1' ),
+			'2.1250' => UnboundedQuantityValue::newFromNumber( '+2.1250', '1' ),
 
-			'1.4e-2' => BoundedQuantityValue::newFromNumber( '+0.014', '1', '+0.015', '+0.013' ),
-			'-1.4e-2' => BoundedQuantityValue::newFromNumber( '-0.014', '1', '-0.013', '-0.015' ),
-			'1.4e3' => BoundedQuantityValue::newFromNumber( '+1400', '1', '+1500', '+1300' ),
-			'1.4e3!m' => BoundedQuantityValue::newFromNumber( '+1400', 'm', '+1400', '+1400' ),
-			'1.4e3m2' => BoundedQuantityValue::newFromNumber( '+1400', 'm2', '+1500', '+1300' ),
-			'1.4ev' => QuantityValue::newFromNumber( '+1.4', 'ev' ),
-			'1.4e' => QuantityValue::newFromNumber( '+1.4', 'e' ),
-			'12e3e4' => BoundedQuantityValue::newFromNumber( '+12000', 'e4', '+13000', '+11000' ),
+			'1.4e-2' => QuantityValue::newFromNumber( '+0.014', '1', '+0.015', '+0.013' ),
+			'-1.4e-2' => QuantityValue::newFromNumber( '-0.014', '1', '-0.013', '-0.015' ),
+			'1.4e3' => QuantityValue::newFromNumber( '+1400', '1', '+1500', '+1300' ),
+			'1.4e3!m' => QuantityValue::newFromNumber( '+1400', 'm', '+1400', '+1400' ),
+			'1.4e3m2' => QuantityValue::newFromNumber( '+1400', 'm2', '+1500', '+1300' ),
+			'1.4ev' => UnboundedQuantityValue::newFromNumber( '+1.4', 'ev' ),
+			'1.4e' => UnboundedQuantityValue::newFromNumber( '+1.4', 'e' ),
+			'12e3e4' => QuantityValue::newFromNumber( '+12000', 'e4', '+13000', '+11000' ),
 			// FIXME: Add support for 12x10^3, see DecimalParser.
-			'0.004e3' => BoundedQuantityValue::newFromNumber( '+4', '1', '+5', '+3' ),
-			'0.004e-3' => BoundedQuantityValue::newFromNumber( '+0.000004', '1', '+0.000005', '+0.000003' ),
-			'4000e3' => BoundedQuantityValue::newFromNumber( '+4000000', '1', '+4001000', '+3999000' ),
-			'4000e-3' => BoundedQuantityValue::newFromNumber( '+4.000', '1', '+4.001', '+3.999' ),
+			'0.004e3' => QuantityValue::newFromNumber( '+4', '1', '+5', '+3' ),
+			'0.004e-3' => QuantityValue::newFromNumber( '+0.000004', '1', '+0.000005', '+0.000003' ),
+			'4000e3' => QuantityValue::newFromNumber( '+4000000', '1', '+4001000', '+3999000' ),
+			'4000e-3' => QuantityValue::newFromNumber( '+4.000', '1', '+4.001', '+3.999' ),
 
 			// precision
-			'0!' => BoundedQuantityValue::newFromNumber( 0, '1', 0, 0 ),
-			'10.003!' => BoundedQuantityValue::newFromNumber( '+10.003', '1', '+10.003', '+10.003' ),
-			'-200!' => BoundedQuantityValue::newFromNumber( -200, '1', -200, -200 ),
-			'0~' => QuantityValue::newFromNumber( 0, '1' ),
-			'10.003~' => QuantityValue::newFromNumber( '+10.003', '1' ),
-			'-200~' => QuantityValue::newFromNumber( -200, '1' ),
+			'0!' => QuantityValue::newFromNumber( 0, '1', 0, 0 ),
+			'10.003!' => QuantityValue::newFromNumber( '+10.003', '1', '+10.003', '+10.003' ),
+			'-200!' => QuantityValue::newFromNumber( -200, '1', -200, -200 ),
+			'0~' => UnboundedQuantityValue::newFromNumber( 0, '1' ),
+			'10.003~' => UnboundedQuantityValue::newFromNumber( '+10.003', '1' ),
+			'-200~' => UnboundedQuantityValue::newFromNumber( -200, '1' ),
 
 			// uncertainty
-			'5.3 +/- 0.2' => BoundedQuantityValue::newFromNumber( '+5.3', '1', '+5.5', '+5.1' ),
-			'5.3+-0.2' => BoundedQuantityValue::newFromNumber( '+5.3', '1', '+5.5', '+5.1' ),
-			'5.3 ±0.2' => BoundedQuantityValue::newFromNumber( '+5.3', '1', '+5.5', '+5.1' ),
+			'5.3 +/- 0.2' => QuantityValue::newFromNumber( '+5.3', '1', '+5.5', '+5.1' ),
+			'5.3+-0.2' => QuantityValue::newFromNumber( '+5.3', '1', '+5.5', '+5.1' ),
+			'5.3 ±0.2' => QuantityValue::newFromNumber( '+5.3', '1', '+5.5', '+5.1' ),
 
-			'5.3 +/- +0.2' => BoundedQuantityValue::newFromNumber( '+5.3', '1', '+5.5', '+5.1' ),
-			'5.3+-+0.2' => BoundedQuantityValue::newFromNumber( '+5.3', '1', '+5.5', '+5.1' ),
+			'5.3 +/- +0.2' => QuantityValue::newFromNumber( '+5.3', '1', '+5.5', '+5.1' ),
+			'5.3+-+0.2' => QuantityValue::newFromNumber( '+5.3', '1', '+5.5', '+5.1' ),
 
-			'5.3e3 +/- 0.2e2' => BoundedQuantityValue::newFromNumber( '+5300', '1', '+5320', '+5280' ),
-			'2e-2+/-1.1e-1' => BoundedQuantityValue::newFromNumber( '+0.02', '1', '+0.13', '-0.09' ),
+			'5.3e3 +/- 0.2e2' => QuantityValue::newFromNumber( '+5300', '1', '+5320', '+5280' ),
+			'2e-2+/-1.1e-1' => QuantityValue::newFromNumber( '+0.02', '1', '+0.13', '-0.09' ),
 
 			// negative
-			'5.3 +/- -0.2' => BoundedQuantityValue::newFromNumber( '+5.3', '1', '+5.5', '+5.1' ),
-			'5.3+--0.2' => BoundedQuantityValue::newFromNumber( '+5.3', '1', '+5.5', '+5.1' ),
-			'5.3 ±-0.2' => BoundedQuantityValue::newFromNumber( '+5.3', '1', '+5.5', '+5.1' ),
+			'5.3 +/- -0.2' => QuantityValue::newFromNumber( '+5.3', '1', '+5.5', '+5.1' ),
+			'5.3+--0.2' => QuantityValue::newFromNumber( '+5.3', '1', '+5.5', '+5.1' ),
+			'5.3 ±-0.2' => QuantityValue::newFromNumber( '+5.3', '1', '+5.5', '+5.1' ),
 
 			// units
-			'5.3+-0.2cm' => BoundedQuantityValue::newFromNumber( '+5.3', 'cm', '+5.5', '+5.1' ),
-			'10.003! km' => BoundedQuantityValue::newFromNumber( '+10.003', 'km', '+10.003', '+10.003' ),
-			'-200~ %  ' => QuantityValue::newFromNumber( -200, '%' ),
-			'100003 m³' => QuantityValue::newFromNumber( 100003, 'm³' ),
-			'3.±-0.2µ' => BoundedQuantityValue::newFromNumber( '+3', 'µ', '+3.2', '+2.8' ),
-			'+00.20 Å' => QuantityValue::newFromNumber( '+0.20', 'Å' ),
+			'5.3+-0.2cm' => QuantityValue::newFromNumber( '+5.3', 'cm', '+5.5', '+5.1' ),
+			'10.003! km' => QuantityValue::newFromNumber( '+10.003', 'km', '+10.003', '+10.003' ),
+			'-200~ %  ' => UnboundedQuantityValue::newFromNumber( -200, '%' ),
+			'100003 m³' => UnboundedQuantityValue::newFromNumber( 100003, 'm³' ),
+			'3.±-0.2µ' => QuantityValue::newFromNumber( '+3', 'µ', '+3.2', '+2.8' ),
+			'+00.20 Å' => UnboundedQuantityValue::newFromNumber( '+0.20', 'Å' ),
 		);
 
 		$argLists = array();
